SDN watch will fix transient errors
Repopulated event queue for every 30 mins.
All the existing items in the queue will be poped as modified events.
We will try to ignore if the item is already processed and previously
failed items will be retried. If the failure is due to transient issue:
temporary network failure to etcd, timing issue, etc. should be fixed.